refactor: CHAT 폴더 및 파일 정리
